+++ src/components/Layout/Layout.tsx
@@ -1,42 +1,13 @@
-import {
-  Center,
-  ColorModeContextType,
-  Flex,
-  HStack,
-  IconButton,
-  Spacer,
-  Tooltip,
-  useColorMode,
-  VStack,
-} from '@chakra-ui/react';
+import { Center, Flex, VStack } from '@chakra-ui/react';
 import React, { FC } from 'react';
 import { Helmet } from 'react-helmet';
-import { useTranslation } from 'react-i18next';
-import {
-  IoArrowBackOutline,
-  IoMoonOutline,
-  IoSunnyOutline,
-} from 'react-icons/io5';
-import {
-  Location,
-  NavigateFunction,
-  useLocation,
-  useNavigate,
-} from 'react-router-dom';
 
 // components
 import Footer from '@app/components/Footer';
+import Header from '@app/components/Header';
 
 // constants
-import {
-  BODY_BACKGROUND_COLOR,
-  DEFAULT_GAP,
-  SEARCH_ROUTE,
-} from '@app/constants';
-
-// hooks
-import useButtonHoverBackgroundColor from '@app/hooks/useButtonHoverBackgroundColor';
-import useDefaultTextColor from '@app/hooks/useDefaultTextColor';
+import { BODY_BACKGROUND_COLOR, DEFAULT_GAP } from '@app/constants';
 
 // selectors
 import { useSelectTitle } from '@app/selectors';
@@ -45,21 +16,8 @@ import { useSelectTitle } from '@app/selectors';
 import type { IProps } from './types';
 
 const Layout: FC<IProps> = ({ children }: IProps) => {
-  const { t } = useTranslation();
-  const location: Location = useLocation();
-  const navigate: NavigateFunction = useNavigate();
   // selectors
   const title: string = useSelectTitle();
-  // hooks
-  const buttonHoverBackgroundColor: string = useButtonHoverBackgroundColor();
-  const { colorMode, toggleColorMode }: ColorModeContextType = useColorMode();
-  const defaultTextColor: string = useDefaultTextColor();
-  // handlers
-  const handlerBackClick = () =>
-    navigate(SEARCH_ROUTE, {
-      replace: true,
-    });
-  const handlerColorChangeClick = () => toggleColorMode();
 
   return (
     <>
@@ -71,50 +29,7 @@ const Layout: FC<IProps> = ({ children }: IProps) => {
         <Flex alignItems="center" justifyContent="center" minH="100vh" w="full">
           <VStack alignItems="center" minH="100vh" spacing={0} w="full">
             {/*header*/}
-            <HStack
-              alignItems="center"
-              justifyContent="space-between"
-              p={DEFAULT_GAP / 2}
-              w="full"
-            >
-              {!location.pathname.includes(SEARCH_ROUTE) && (
-                <Tooltip label={t('captions.goBackToSearch')}>
-                  <IconButton
-                    _hover={{ bg: buttonHoverBackgroundColor }}
-                    aria-label="Go back"
-                    color={defaultTextColor}
-                    icon={<IoArrowBackOutline />}
-                    onClick={handlerBackClick}
-                    size="lg"
-                    variant="ghost"
-                  />
-                </Tooltip>
-              )}
-
-              <Spacer />
-
-              <Tooltip
-                label={t('captions.switchColorMode', {
-                  colorMode: colorMode === 'dark' ? 'light' : 'dark',
-                })}
-              >
-                <IconButton
-                  _hover={{ bg: buttonHoverBackgroundColor }}
-                  aria-label="Change color mode"
-                  color={defaultTextColor}
-                  icon={
-                    colorMode === 'dark' ? (
-                      <IoSunnyOutline />
-                    ) : (
-                      <IoMoonOutline />
-                    )
-                  }
-                  onClick={handlerColorChangeClick}
-                  size="lg"
-                  variant="ghost"
-                />
-              </Tooltip>
-            </HStack>
+            <Header />
 
             {/*content*/}
             <VStack
